+++ docs/versionhistory.rst
@@ -5,6 +5,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 **1.4.0**
 
+- Added async name resolution functions (``anyio.getaddrinfo()`` and ``anyio.getnameinfo()``)
 - Fixed ``fail.after(0)`` not raising a timeout error on asyncio and curio
 - Fixed ``move_on_after()`` and ``fail_after()`` getting stuck on curio in some circumstances
 - Fixed socket operations not allowing timeouts to cancel the task
@@ -15,6 +16,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
   on asyncio and curio
 - Fixed the type of the ``address`` parameter in ``UDPSocket.send()`` to include ``IPAddress``
   objects (which were already supported by the backing implementation)
+- Switched to using ``anyio.getaddrinfo()`` for name lookups
 
 **1.3.1**